+++ src/main/kotlin/com/atlassian/performance/tools/aws/api/StackFormula.kt
@@ -1,9 +1,6 @@
 package com.atlassian.performance.tools.aws.api
 
-import com.amazonaws.services.cloudformation.model.Capability
-import com.amazonaws.services.cloudformation.model.CreateStackRequest
-import com.amazonaws.services.cloudformation.model.Parameter
-import com.amazonaws.services.cloudformation.model.Stack
+import com.amazonaws.services.cloudformation.model.*
 import com.amazonaws.services.cloudformation.model.StackStatus.*
 import org.apache.commons.codec.binary.Hex
 import org.apache.logging.log4j.LogManager
@@ -112,10 +109,29 @@ data class StackFormula @JvmOverloads constructor(
                 else -> {
                     logger.error("Stack $stack failed: $status")
                     aws.stackNanny.takeCare(stackName)
-                    throw Exception("Stack $stackName creation failed: $stack")
+                    throwCreationFailureReason(stack)
                 }
             }
         }
         throw Exception("Stack $stackName provisioning timed out: $pollingTimeout")
+    }
+
+    private fun throwCreationFailureReason(stack: Stack) {
+        try {
+            val stackCreationReason = aws.cloudformation
+                .describeStackEvents(DescribeStackEventsRequest().withStackName(stack.stackId))
+                .stackEvents
+                .filter { ResourceStatus.fromValue(it.resourceStatus) == ResourceStatus.CREATE_FAILED }
+                .takeIf { it.isNotEmpty() }
+                ?.joinToString("; ") { "${it.logicalResourceId} failed to create because: ${it.resourceStatusReason}" }
+
+            if (stackCreationReason != null) {
+                throw Exception("Stack $stackName creation failed, because <<${stackCreationReason}>>: $stack")
+            } else {
+                throw Exception("Stack $stackName creation failed, but we didn't find any related event with status CREATE_FAILED: $stack")
+            }
+        } catch (e: Exception) {
+            throw Exception("Stack $stackName creation failed, but we failed to capture failure reason: $stack", e)
+        }
     }
 }