+++ CHANGELOG.md
@@ -23,6 +23,11 @@ Dropping a requirement of a major version of a dependency is a new contract.
 ## [Unreleased]
 [Unreleased]: https://bitbucket.org/atlassian/aws-resources/branches/compare/master%0Drelease-1.8.1
 
+### Fixed
+- Capture failure reason of stack creation as part of `StackFormula`. Resolves [JPERF-494].
+
+[JPERF-494]: https://ecosystem.atlassian.net/browse/JPERF-494
+
 ## [1.8.1] - 2022-02-16
 [1.8.1]: https://bitbucket.org/atlassian/aws-resources/branches/compare/release-1.8.1%0Drelease-1.8.0
 
@@ -35,6 +40,8 @@ Dropping a requirement of a major version of a dependency is a new contract.
 ### Added
 - `SshInstance` now contains AWS SDK `Instance` object, which includes i.a. information about private IP address. Unblocks [JPERF-730].
 - `ProvisionedStack` now allows to find its security groups based on provided logical id. Unblocks [JPERF-730].
+
+[JPERF-730]: https://ecosystem.atlassian.net/browse/JPERF-730
 
 ## [1.7.1] - 2021-09-28
 [1.7.1]: https://bitbucket.org/atlassian/aws-resources/branches/compare/release-1.7.1%0Drelease-1.7.0